# openapi/openapi.yaml
openapi: 3.1.0
info:
  description: "Online task planning system — manage tasks, work slots, dependencies\
    \ and generate optimized plans."
  title: ChronoScope API
  version: 1.1.0
servers:
- description: Production
  url: https://chronoscope.ni0.team/api
security:
- oauth2: []
...
paths:
  /v1/identity:
    get:
      description: "Retrieve information about the authenticated identity, including\
        \ all linked accounts and the organizations for which the identity has admin\
        \ privileges, as resolved server-side."
      operationId: getIdentity
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/IdentityResponse"
          description: Identity retrieved successfully
        "401":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Missing or invalid token
      summary: Get current identity
      tags:
      - Identity
...
  /v1/workslots:
    get:
      description: Return all work slots belonging to the current identity.
      operationId: getWorkSlots
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkSlotResponse"
          description: Work slots retrieved successfully
        "401":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Missing or invalid token
      summary: List work slots
      tags:
      - Work Slots
    post:
      description: Create a new available work slot for the given account.
      operationId: createWorkSlot
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkSlotCreateRequest"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WorkSlotResponse"
          description: Work slot created
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Validation error
        "401":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Missing or invalid token
      summary: Create work slot
      tags:
      - Work Slots
  /v1/workslots/{id}:
    delete:
      description: Delete a work slot.
      operationId: deleteWorkSlot
      parameters:
      - description: Work slot ID
        in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Work slot deleted
        "401":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Missing or invalid token
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Work slot not found
      summary: Delete work slot
      tags:
      - Work Slots
    patch:
      description: Update the start/end times of an existing work slot (PATCH semantics).
      operationId: updateWorkSlot
      parameters:
      - description: Work slot ID
        in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkSlotUpdateRequest"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WorkSlotResponse"
          description: Work slot updated
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Validation error
        "401":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Missing or invalid token
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Work slot not found
      summary: Update work slot
      tags:
      - Work Slots
components:
  schemas:
...
    AccountResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        identityId:
          type: integer
          format: int64
      required:
      - id
      - identityId
...
    IdentityResponse:
      type: object
      properties:
        accounts:
          type: array
          items:
            $ref: "#/components/schemas/AccountResponse"
        adminOrganizations:
          type: array
          description: "Organization IDs for which the authenticated identity has\
            \ admin privileges, resolved server-side via Keycloak Admin Client lookups."
          items:
            type: string
          uniqueItems: true
        id:
          type: integer
          format: int64
        organizations:
          type: array
          description: Organizations to which the identity belongs.
          items:
            $ref: "#/components/schemas/Organization"
          uniqueItems: true
      required:
      - accounts
      - adminOrganizations
      - id
      - organizations
...
    Organization:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
...
    ProblemDetail:
      type: object
      properties:
        detail:
          type: string
        instance:
          type: string
          format: uri
        properties:
          type: object
          additionalProperties: {}
        status:
          type: integer
          format: int32
        title:
          type: string
        type:
          type: string
          format: uri
...
    WorkSlotCreateRequest:
      type: object
      properties:
        endAt:
          type: string
          format: date-time
        organizationId:
          type: string
        startAt:
          type: string
          format: date-time
      required:
      - endAt
      - organizationId
      - startAt
    WorkSlotResponse:
      type: object
      properties:
        endAt:
          type: string
          format: date-time
        id:
          type: integer
          format: int64
        organizationId:
          type: string
        startAt:
          type: string
          format: date-time
      required:
      - endAt
      - id
      - organizationId
      - startAt
    WorkSlotUpdateRequest:
      type: object
      properties:
        endAt:
          type: string
          format: date-time
        startAt:
          type: string
          format: date-time
  securitySchemes:
    oauth2:
      description: Keycloak OAuth2 (Authorization Code flow)
      flows:
        authorizationCode:
          authorizationUrl: https://auth.ni0.team/realms/ni0/protocol/openid-connect/auth
          scopes:
            read: Read access
            write: Write access
          tokenUrl: https://auth.ni0.team/realms/ni0/protocol/openid-connect/token
      type: oauth2